+++ src/main/java/guy/storage/Storage.java
@@ -1,9 +1,11 @@
-package storage;
+package guy.storage;
 
 import java.io.*;
 import java.time.format.DateTimeFormatter;
 import java.util.*;
 
+import guy.tasks.Task;
+import guy.tasks.TaskManager;
 import tasks.*;
 
 public class Storage {
