+++ src/main/java/guy/ui/Ui.java
@@ -1,0 +1,14 @@
+package guy.ui;
+
+public class Ui {
+    String name = "that one guy";
+
+    public void greet() {
+        System.out.println("I'm " + name + ".");
+        System.out.println("Make it quick, I don't have much time.");
+    }
+
+    public void bye() {
+        System.out.println("Whatever. Hope you never come back.");
+    }
+}
