+++ src/test/java/guy/ThatOneGuyTest.java
@@ -1,0 +1,2 @@
+public class ThatOneGuyTest {
+}
